+++ src/index.ts
@@ -28,3 +28,4 @@ export { ReactionMessage } from './messages/reaction-message';
 export { PollCreationMessage, PollOption } from './messages/poll-creation-message';
 export { ButtonsMessage, Button } from './messages/buttons-message';
 export { ListMessage, Section, Row } from './messages/list-message';
+export { ProtocolMessage, ProtocolMessageType } from './messages/protocol-message';